+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/compat/utils/__init__.py
@@ -1,0 +1,11 @@
+# Copyright (c) 2022, NVIDIA CORPORATION & AFFILIATES, ETH Zurich, and University of Toronto
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+from .configclass import configclass
+
+__all__ = [
+    # config wrapper
+    "configclass",
+]