+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/utils/io/yaml.py
@@ -31,7 +31,7 @@ def load_yaml(filename: str) -> Dict:
     return data
 
 
-def dump_yaml(filename: str, data: Union[Dict, object]):
+def dump_yaml(filename: str, data: Union[Dict, object], sort_keys: bool = False):
     """Saves data into a YAML file safely.
 
     Note:
@@ -40,6 +40,7 @@ def dump_yaml(filename: str, data: Union[Dict, object]):
     Args:
         filename (str): The path to save the file at.
         data (Union[Dict, object]): The data to save either a dictionary or class object.
+        sort_keys (bool, optional): Whether to sort the keys in the output file. Defaults to False.
     """
     # check ending
     if not filename.endswith("yaml"):
@@ -52,4 +53,4 @@ def dump_yaml(filename: str, data: Union[Dict, object]):
         data = class_to_dict(data)
     # save data
     with open(filename, "w") as f:
-        yaml.dump(data, f, default_flow_style=None)
+        yaml.dump(data, f, default_flow_style=None, sort_keys=sort_keys)